"ads2_m12_a1_1: 0/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ M 12/ADS-2_Exam-2/EdgeWeightedGraph.java
@@ -181,7 +181,7 @@ public class EdgeWeightedGraph {
             s.append(v + ": ");
             for (Edge e : adj[v]) {
                 s.append(e.either() + "-" + e.other(e.either()) +
-                    " " + e.weight() + "  ");
+                    " " + (double) e.weight() + "  ");
             }
             s.append(NEWLINE);
         }
